Refactored app/lib/* to lib/*
